Improved product, shoppinglist and user, added user tests

+++ src/main/java/com/ajcompare/service/ShoppingListProductService.java
@@ -1,17 +1,17 @@
 package com.ajcompare.service;
 
-import com.ajcompare.domain.Product;
 import com.ajcompare.domain.ShoppingListProduct;
-import com.ajcompare.repository.ProductRepository;
+import com.ajcompare.domain.User;
 import com.ajcompare.repository.ShoppingListProductRepository;
+import com.ajcompare.repository.UserRepository;
 import io.quarkus.panache.common.Parameters;
-import io.quarkus.panache.common.Sort;
+import io.quarkus.security.UnauthorizedException;
+import io.quarkus.security.identity.SecurityIdentity;
 
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 import javax.transaction.Transactional;
 import javax.ws.rs.NotFoundException;
-import java.util.ArrayList;
 import java.util.List;
 
 @ApplicationScoped
@@ -19,6 +19,12 @@ public class ShoppingListProductService {
 
     @Inject
     ShoppingListProductRepository shoppingListProductRepository;
+
+    @Inject
+    UserRepository userRepository;
+
+    @Inject
+    SecurityIdentity securityIdentity;
 
     public ShoppingListProductService() {
     }
@@ -40,17 +46,48 @@ public class ShoppingListProductService {
 
 
     public List<ShoppingListProduct> allShoppingListProducts(Integer userId) {
+        if (userId == null)
+        {
+            throw new IllegalArgumentException();
+        }
+        User user = userRepository.find("id", userId).firstResult();
+
+        if (!securityIdentity.getPrincipal().getName().equals(user.getName())){
+            throw new UnauthorizedException();
+        }
+
         Integer shoppingListId = getLastShoppingListId(userId);
 
         return shoppingListProductRepository.list("from ShoppingListProduct where shoppingListId = ?1 and userId = ?2", shoppingListId, userId);
     }
 
     public List<ShoppingListProduct> getOldShoppingListProducts(Integer userId, Integer shoppingListId) {
+        if (userId == null || shoppingListId == null)
+        {
+            throw new IllegalArgumentException();
+        }
+        ShoppingListProduct shoppingListProduct = shoppingListProductRepository.find("shoppingListId", shoppingListId).firstResult();
+        User user = userRepository.find("id", shoppingListProduct.getUserId()).firstResult();
+
+        if (!securityIdentity.getPrincipal().getName().equals(user.getName())){
+            throw new UnauthorizedException();
+        }
+
         return shoppingListProductRepository.list("from ShoppingListProduct where shoppingListId = ?1 and userId = ?2", shoppingListId, userId);
     }
 
     public ShoppingListProduct getShoppingListProductById(Integer productId) {
+        if (productId == null)
+        {
+            throw new IllegalArgumentException();
+        }
         ShoppingListProduct shoppingListProduct = shoppingListProductRepository.find("id", productId).firstResult();
+        User user = userRepository.find("id", shoppingListProduct.getUserId()).firstResult();
+
+        if (!securityIdentity.getPrincipal().getName().equals(user.getName())){
+            throw new UnauthorizedException();
+        }
+
         if (shoppingListProduct == null)
         {
             throw new NotFoundException();
@@ -60,14 +97,29 @@ public class ShoppingListProductService {
 
     @Transactional
     public Long deleteShoppingListProduct(Integer productId) {
+        if (productId == null)
+        {
+            throw new IllegalArgumentException();
+        }
+        ShoppingListProduct shoppingListProduct = shoppingListProductRepository.find("id", productId).firstResult();
+        User user = userRepository.find("id", shoppingListProduct.getUserId()).firstResult();
+
+        if (!securityIdentity.getPrincipal().getName().equals(user.getName())){
+            throw new UnauthorizedException();
+        }
         return shoppingListProductRepository.delete("id", productId);
     }
 
     @Transactional
     public ShoppingListProduct updateShoppingListProduct(ShoppingListProduct shoppingListProduct) {
+        User user = userRepository.find("id", shoppingListProduct.getUserId()).firstResult();
+
+        if (!securityIdentity.getPrincipal().getName().equals(user.getName())){
+            throw new UnauthorizedException();
+        }
 
         ShoppingListProduct shoppingListProductTemp = shoppingListProductRepository.find("id", shoppingListProduct.getId()).firstResult();
-        if (shoppingListProduct.getSuperMarket() != ""){
+        if (shoppingListProduct.getSuperMarket() != null){
             shoppingListProductTemp.setSuperMarket(shoppingListProduct.getSuperMarket());
         }
         if (shoppingListProduct.getContent() != null){
@@ -76,30 +128,21 @@ public class ShoppingListProductService {
         if (shoppingListProduct.getQuantity() != null){
             shoppingListProductTemp.setQuantity(shoppingListProduct.getQuantity());
         }
-        if (shoppingListProduct.getName() != ""){
+        if (shoppingListProduct.getName() != null){
             shoppingListProductTemp.setName(shoppingListProduct.getName());
         }
-        if (shoppingListProduct.getUrl() != ""){
+        if (shoppingListProduct.getUrl() != null){
             shoppingListProductTemp.setUrl(shoppingListProduct.getUrl());
         }
         if (shoppingListProduct.getDate() != null){
             shoppingListProductTemp.setDate(shoppingListProduct.getDate());
         }
-        if (shoppingListProduct.getPrice() != null   ){
+        if (shoppingListProduct.getPrice() != null){
             shoppingListProductTemp.setPrice(shoppingListProduct.getPrice());
         }
-        if (shoppingListProduct.getSuperMarket() != null   ){
+        if (shoppingListProduct.getSuperMarket() != null){
             shoppingListProductTemp.setSuperMarket(shoppingListProduct.getSuperMarket());
         }
-//        shoppingListProductRepository.update("update ShoppingListProduct set name = :name, quantity= :quantity, " +
-//                        "content = :content,  url = :url, date = :date, " +
-//                        "price = :price where id = :id",)
-//
-//                new Parameters().with("name", shoppingListProductTemp.getName()).and("quantity", shoppingListProductTemp.getQuantity()).and(
-//                "content", shoppingListProductTemp.getContent()).and("url", shoppingListProductTemp.getUrl()).and(
-//                        "date", shoppingListProductTemp.getDate()).and("price", shoppingListProductTemp.getPrice()).and(
-//                                "id", shoppingListProductTemp.getId());
-
 
         shoppingListProductRepository.update("update ShoppingListProduct set name = :name, quantity= :quantity, content = :content,  url = :url, date = :date, price = :price where id = :id",
                 new Parameters().with("name", shoppingListProductTemp.getName()).and("quantity", shoppingListProductTemp.getQuantity()).and("content", shoppingListProductTemp.getContent()).and("url", shoppingListProductTemp.getUrl()).and("date", shoppingListProductTemp.getDate()).and("price", shoppingListProductTemp.getPrice()).and("id", shoppingListProductTemp.getId()));
@@ -109,9 +152,15 @@ public class ShoppingListProductService {
 
     @Transactional
     public ShoppingListProduct addShoppingListProduct(Integer userId, ShoppingListProduct shoppingListProduct) {
-        if (shoppingListProduct == null)
+        if (shoppingListProduct == null || userId == null)
         {
             throw new IllegalArgumentException();
+        }
+
+        User user = userRepository.find("id", userId).firstResult();
+
+        if (!securityIdentity.getPrincipal().getName().equals(user.getName())){
+            throw new UnauthorizedException();
         }
 
         Integer shoppingListId = getLastShoppingListId(userId);
